+++ microlens-mtl/CHANGELOG.md
@@ -1,3 +1,7 @@
+# next
+
+* Exported `Focusing`, etc. from `Lens.Micro.Mtl.Internal`.
+
 # 0.1.10.0
 
 * Added `<?=` and `<.=`.